edit and delete buttons are now greyed out for users who are not owners
they cant also use the url to access those pages

+++ src/components/DeleteCar.js
@@ -7,15 +7,21 @@ import { AuthContext } from '../contexts/AuthContext';
 const DeleteCar = ({ cars, deleteCarHandler }) => {
     const { carId } = useParams();
     const [car, setCar] = useState({});
-    const {user} = useContext(AuthContext);
+    const { user } = useContext(AuthContext);
     const navigate = useNavigate();
 
     useEffect(() => {
         setCar(cars.find(x => x._id === carId));
     }, [carId]);
 
+    const carBelongsToUser = car._ownerId === user._id;
+
     const onClick = (ev) => {
         ev.preventDefault();
+
+        if (!carBelongsToUser) {
+            return;
+        }
 
         carService.deleteCar(carId);
         deleteCarHandler(carId);
